requirements.txt
clipspy
tk
